Fix loss of precision in OpenCvSharp.Size2d constructor

+++ src/OpenCvSharp/Modules/core/Struct/Size2d.cs
@@ -38,8 +38,8 @@ namespace OpenCvSharp
         /// <param name="height"></param>
         public Size2d(double width, double height)
         {
-            Width = (int) width;
-            Height = (int) height;
+            Width = width;
+            Height = height;
         }
 
         #region Operators
